no_and_co/requirements.txt
arabic-reshaper==3.0.1
asgiref==3.11.1
asn1crypto==1.5.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cloudinary==1.44.2
cryptography==47.0.0
cssselect2==0.9.0
Django==5.2.13
django-allauth==65.15.1
django-cloudinary-storage==0.3.0
freetype-py==2.5.1
gunicorn==25.3.0
html5lib==1.1
idna==3.13
lxml==6.1.0
oscrypto==1.3.0
packaging==26.2
phonenumbers==9.0.27
pillow==12.2.0
psycopg2-binary==2.9.12
pycairo==1.29.0
pycparser==3.0
pyHanko==0.34.1
pyhanko-certvalidator==0.30.2
PyJWT==2.12.1
pypdf==6.10.2
python-bidi==0.6.7
python-decouple==3.8
PyYAML==6.0.3
razorpay==2.0.1
reportlab==4.5.0
requests==2.33.1
rlPyCairo==0.4.0
six==1.17.0
sqlparse==0.5.5
svglib==1.6.0
tinycss2==1.5.1
tzlocal==5.3.1
uritools==6.1.0
urllib3==2.6.3
webencodings==0.5.1
xhtml2pdf==0.2.17
